requirements.txt
# Core Data Handling and Numerics
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning and Modeling
scikit-learn>=1.2.0
xgboost>=1.7.0
joblib>=1.2.0

# Data Ingestion
yfinance>=0.2.0
requests>=2.28.0
fredapi>=0.5.0

# Dashboard and Visualization
plotly>=5.14.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0
shap>=0.41.0

# Environment and Configuration Loader
python-dotenv>=1.0.0
matplotlib>=3.7.0